ldap sssd libsss-sudo, bash changes, ahenk post install sssd_ad_authentication

+++ src/base/command/command_runner.py
@@ -73,7 +73,7 @@ class CommandRunner(object):
 
                 if str(json_data['event']) == 'login' and self.check_last_login():
                     username = json_data['username']
-                    if username != "Debian-gdm" or username != "gdm":
+                    if username != "Debian-gdm" and username != "gdm":
                         display = json_data['display']
                         desktop = json_data['desktop']
 
@@ -163,22 +163,23 @@ class CommandRunner(object):
 
                 elif str(json_data['event']) == 'logout':
                     username = json_data['username']
-                    self.execute_manager.remove_user_executed_policy_dict(username)
-                    self.plugin_manager.process_mode('logout', username)
-                    self.plugin_manager.process_mode('safe', username)
-                    if username != "Debian-gdm":
-                        self.db_service.delete('session', '1=1')
-                    # TODO delete all user records while initializing
-                    self.logger.info('logout event is handled for user: {0}'.format(username))
-                    ip = None
-                    if 'ip' in json_data:
-                        ip = json_data['ip']
+                    if username != "Debian-gdm" and username != "gdm":
+                        self.execute_manager.remove_user_executed_policy_dict(username)
+                        self.plugin_manager.process_mode('logout', username)
+                        self.plugin_manager.process_mode('safe', username)
+                        if username != "Debian-gdm":
+                            self.db_service.delete('session', '1=1')
+                        # TODO delete all user records while initializing
+                        self.logger.info('logout event is handled for user: {0}'.format(username))
+                        ip = None
+                        if 'ip' in json_data:
+                            ip = json_data['ip']
 
-                    logout_message = self.message_manager.logout_msg(username,ip)
-                    self.messenger.send_direct_message(logout_message)
-                    self.logger.info('Ahenk polkit file deleting..')
-                    self.delete_polkit_user()
-                    # self.db_service.delete('session', 'username=\'{0}\''.format(username))
+                        logout_message = self.message_manager.logout_msg(username,ip)
+                        self.messenger.send_direct_message(logout_message)
+                        self.logger.info('Ahenk polkit file deleting..')
+                        self.delete_polkit_user()
+                        # self.db_service.delete('session', 'username=\'{0}\''.format(username))
 
                 elif str(json_data['event']) == 'send':
                     self.logger.info('Sending message over ahenkd command. Response Message: {0}'.format(
